extend timeout, refactor out huge string blob
