+++ vitest/runtime/custom-elements/custom-elements.test.js
@@ -11,101 +11,106 @@ const index = path.resolve('index.mjs');
 
 const browser_assert = fs.readFileSync(`${__dirname}/assert.js`, 'utf-8');
 
-describe('custom-elements', function () {
-	/** @type {import('@playwright/test').Browser} */
-	let browser;
-	beforeAll(async () => {
-		browser = await chromium.launch();
-		console.log('[custom-elements] Launched browser');
-	}, 20000);
+describe(
+	'custom-elements',
+	() => {
+		/** @type {import('@playwright/test').Browser} */
+		let browser;
 
-	afterAll(async () => {
-		if (browser) await browser.close();
-	});
+		beforeAll(async () => {
+			browser = await chromium.launch();
+			console.log('[custom-elements] Launched browser');
+		}, 20000);
 
-	fs.readdirSync(`${__dirname}/samples`).forEach((dir) => {
-		if (dir[0] === '.') return;
+		afterAll(async () => {
+			if (browser) await browser.close();
+		});
 
-		const solo = /\.solo$/.test(dir);
-		const skip = /\.skip$/.test(dir);
+		fs.readdirSync(`${__dirname}/samples`).forEach((dir) => {
+			if (dir[0] === '.') return;
 
-		const warnings = [];
-		const it_fn = solo ? it.only : skip ? it.skip : it;
+			const solo = /\.solo$/.test(dir);
+			const skip = /\.skip$/.test(dir);
 
-		it_fn(dir, async () => {
-			const config = await try_load_module(import(`./samples/${dir}/_config.js`));
+			const warnings = [];
+			const it_fn = solo ? it.only : skip ? it.skip : it;
 
-			const expected_warnings = config.warnings || [];
+			it_fn(dir, async () => {
+				const config = await try_load_module(import(`./samples/${dir}/_config.js`));
 
-			const bundle = await rollup({
-				input: `${__dirname}/samples/${dir}/test.js`,
-				plugins: [
-					// @ts-ignore -- TODO: fix this
-					{
-						resolveId(importee) {
-							if (importee === 'svelte/internal' || importee === './internal') {
-								return internal;
-							}
+				const expected_warnings = config.warnings || [];
 
-							if (importee === 'svelte') {
-								return index;
-							}
+				const bundle = await rollup({
+					input: `${__dirname}/samples/${dir}/test.js`,
+					plugins: [
+						// @ts-ignore -- TODO: fix this
+						{
+							resolveId(importee) {
+								if (importee === 'svelte/internal' || importee === './internal') {
+									return internal;
+								}
 
-							if (importee === 'assert') {
-								return 'assert';
-							}
-						},
+								if (importee === 'svelte') {
+									return index;
+								}
 
-						load(id) {
-							if (id === 'assert') return browser_assert;
-						},
+								if (importee === 'assert') {
+									return 'assert';
+								}
+							},
 
-						transform(code, id) {
-							if (id.endsWith('.svelte')) {
-								const compiled = svelte.compile(code.replace(/\r/g, ''), {
-									customElement: true,
-									dev: config.dev
-								});
+							load(id) {
+								if (id === 'assert') return browser_assert;
+							},
 
-								compiled.warnings.forEach((w) => warnings.push(w));
+							transform(code, id) {
+								if (id.endsWith('.svelte')) {
+									const compiled = svelte.compile(code.replace(/\r/g, ''), {
+										customElement: true,
+										dev: config.dev
+									});
 
-								return compiled.js;
+									compiled.warnings.forEach((w) => warnings.push(w));
+
+									return compiled.js;
+								}
 							}
 						}
+					]
+				});
+
+				const generated_bundle = await bundle.generate({ format: 'iife', name: 'test' });
+
+				function assertWarnings() {
+					if (expected_warnings) {
+						assert.deepStrictEqual(
+							warnings.map((w) => ({
+								code: w.code,
+								message: w.message,
+								pos: w.pos,
+								start: w.start,
+								end: w.end
+							})),
+							expected_warnings
+						);
 					}
-				]
-			});
-
-			const generated_bundle = await bundle.generate({ format: 'iife', name: 'test' });
-
-			function assertWarnings() {
-				if (expected_warnings) {
-					assert.deepStrictEqual(
-						warnings.map((w) => ({
-							code: w.code,
-							message: w.message,
-							pos: w.pos,
-							start: w.start,
-							end: w.end
-						})),
-						expected_warnings
-					);
 				}
-			}
 
-			const page = await browser.newPage();
-			page.on('console', (type) => {
-				console[type.type()](type.text());
+				const page = await browser.newPage();
+				page.on('console', (type) => {
+					console[type.type()](type.text());
+				});
+				await page.setContent('<main></main>');
+				await page.evaluate(generated_bundle.output[0].code);
+				const test_result = await page.evaluate(`test(document.querySelector('main'))`);
+
+				if (test_result) console.log(test_result);
+
+				assertWarnings();
+
+				await page.close();
 			});
-			await page.setContent('<main></main>');
-			await page.evaluate(generated_bundle.output[0].code);
-			const test_result = await page.evaluate(`test(document.querySelector('main'))`);
-
-			if (test_result) console.log(test_result);
-
-			assertWarnings();
-
-			await page.close();
 		});
-	});
-});
+	},
+	{ timeout: 20000 }
+);
